data-pipeline/stream_emitting/requirements.txt
kafka-python==2.0.2
pandas=1.4.4